requirements.txt
google-cloud-pubsub
requests
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2
google-cloud-secret-manager